+++ src/module/vue/components/site/site-movebox.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="box flexrow clickable block" @click="click">
+  <div class="box flexrow clickable block" :class="{disabled: disabled}" @click="click">
     <h4 class="nogrow" style="margin: 0; white-space: nowrap">
       {{ $t(i18nKey) }}
     </h4>
@@ -11,13 +11,10 @@ export default {
   props: {
     actor: Object,
     move: String,
+    disabled: Boolean,
   },
 
   computed: {
-    ironswornActor() {
-      return game.actors?.get(this.actor._id)
-    },
-
     i18nKey() {
       return `IRONSWORN.MoveContents.${this.move}.title`
     },
@@ -25,8 +22,9 @@ export default {
 
   methods: {
     async click() {
+      if (this.disabled) return
       const move = await CONFIG.IRONSWORN.moveDataByName(this.move)
-      CONFIG.IRONSWORN.RollDialog.show({ move, site: this.ironswornActor })
+      CONFIG.IRONSWORN.RollDialog.show({ move, site: this.$actor() })
     },
   },
 }